pyproject.toml
[tool.poetry]
name = "appjsonify"
version = "0.1.0"
description = "An academic paper PDF to JSON conversion toolkit."
authors = ["Atsuki Yamaguchi <atsuki.yamaguchi.xn@hitachi.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
registrable = "^0.0.4"
pdfplumber = "^0.10.2"
tqdm = "^4.65.0"
pillow = "^10.0.0"
pdf2image = "^1.16.3"
torch = {version = "^2.0.1+cu118", source = "torch_cu118"}
torchvision = {version = "^0.15.2+cu118", source = "torch_cu118"}
torchaudio = {version = "^2.0.2+cu118", source = "torch_cu118"}
opencv-python = "^4.8.0.76"

[tool.poetry.scripts]
appjsonify = "appjsonify.main:run_cli"

[[tool.poetry.source]]
name = "torch_cu118"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[tool.coverage.run]
omit = [
    "appjsonify/modules/load/models/detectron2_demo/*",
]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
